Remove the dead history_loader module
`history_loader::load_all` was deleted when host history moved to ACP
`session/list`, but the module stayed behind: half of it was disk-scanner
dead code kept compiling only by `#[allow(dead_code)]`, and its header
still documented a per-CLI on-disk scan that no longer runs. The name
alone implied WTA cold-scans four CLI transcript stores at startup.

Dissolve it into the code that actually uses the survivors:

- `parse_iso_to_system_time` and `short_id` move to `session_history.rs`
  next to their only callers (ACP `SessionInfo::updated_at` and the
  no-title fallback label), with the Codex-specific doc rewritten.
- `codex_record_is_subagent_meta` / `codex_payload_is_subagent` move to
  `session_watcher/classify_codex.rs` as `record_is_subagent_meta` /
  `payload_is_subagent`, beside the classifier that consumes them.
- Delete the YAML block-scalar parser, the JSONL title extractors and
  their helpers, all unused.

Also refresh comments that still pointed at `load_all` or the removed
`wsl_acp` module, and mark the two spec sections describing the deleted
disk scanner as superseded.

No behavior change; tests move with their code.

Copilot-Session: 741e2eec-21f3-4286-8e2e-500eb06d553d
