+++ tools/wta/src/session_watcher/classify_codex.rs
@@ -94,6 +94,26 @@ fn codex_permission_reason(payload: Option<&serde_json::Value>) -> Option<String
     Some(reason)
 }
 
+/// True if a Codex rollout record is the `session_meta` of an internal
+/// multi-agent subagent / forked thread. Codex's `multi_agent_v1` / `spawn_agent`
+/// tool forks a child thread that gets its own `rollout-*.jsonl` (carrying
+/// `source.subagent` in its meta) and inherits the parent's full history — so it
+/// shows the same first user message / title. It is a codex-internal worker, not
+/// a user-facing session, and must not surface as its own session row.
+pub fn record_is_subagent_meta(v: &serde_json::Value) -> bool {
+    v.get("type").and_then(|t| t.as_str()) == Some("session_meta")
+        && v.get("payload").map(payload_is_subagent).unwrap_or(false)
+}
+
+/// True if a Codex `session_meta` payload's `source` is the subagent variant
+/// (`{"subagent": …}`) rather than a top-level session (`"cli"` / `"user"`).
+pub fn payload_is_subagent(payload: &serde_json::Value) -> bool {
+    payload
+        .get("source")
+        .and_then(|s| s.get("subagent"))
+        .is_some()
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -198,5 +218,20 @@ mod tests {
     fn plain_message_yields_nothing() {
         let r = rec(r#"{"type":"response_item","payload":{"type":"message","role":"user"}}"#);
         assert!(classify(&r, &"k".to_string()).is_empty());
+    }
+
+    #[test]
+    fn payload_is_subagent_discriminates_source() {
+        let cli = serde_json::json!({ "source": "cli" });
+        assert!(
+            !payload_is_subagent(&cli),
+            "top-level source=\"cli\" is not a subagent"
+        );
+        let sub =
+            serde_json::json!({ "source": { "subagent": { "thread_spawn": { "depth": 1 } } } });
+        assert!(
+            payload_is_subagent(&sub),
+            "source.subagent must be detected"
+        );
     }
 }
